requirements.txt
flask
numpy
tensorflow
google-genai
python-dotenv
supabase
